requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
WTForms==3.0.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
python-telegram-bot==20.6
APScheduler==3.10.4
phonenumbers==8.13.19
email-validator==2.0.0
Pillow==10.0.1
lxml==4.9.3
defusedxml==0.7.1
cloudpayments==1.3.0
reportlab==4.0.4
PyJWT==2.8.0